Participation Controller 분리, 참여 탈퇴 추가
